requirements.txt
Flask
azure-ai-translation-text
azure-ai-vision-imageanalysis
python-dotenv
requests